Changed the Class component to Function Component
